Move metrics out of the http module
