+++ build.sbt
@@ -97,6 +97,7 @@ lazy val libats_http = (project in file("libats-http"))
   .settings(commonSettings)
   .settings(libraryDependencies ++= Library.akkaHttp)
   .settings(libraryDependencies ++= Library.jvmMetrics)
+  .settings(libraryDependencies ++= Library.circe)
   .settings(libraryDependencies ++= Library.akkaHttpTestKit)
   .settings(Publish.settings)
   .dependsOn(libats)
@@ -140,6 +141,7 @@ lazy val libats_messaging = (project in file("libats-messaging"))
   .settings(Publish.settings)
   .settings(libraryDependencies ++= Library.akkaHttpTestKit)
   .dependsOn(libats)
+  .dependsOn(libats_metrics)
   .dependsOn(libats_http)
   .dependsOn(libats_messaging_datatype)
