+++ libats-metrics/src/main/scala/com/advancedtelematic/metrics/MetricsRepresentation.scala
@@ -1,0 +1,11 @@
+package com.advancedtelematic.metrics
+
+import io.circe.Json
+
+import scala.concurrent.Future
+
+trait MetricsRepresentation {
+  def metricsJson: Future[Json]
+
+  def urlPrefix: String
+}
