+++ libats-slick/src/main/scala/com/advancedtelematic/libats/slick/monitoring/DatabaseMetrics.scala
@@ -2,10 +2,10 @@ package com.advancedtelematic.libats.slick.monitoring
 
 import java.util
 import java.lang.management.ManagementFactory
-import javax.management.{JMX, ObjectName}
 
-import com.advancedtelematic.libats.http.monitoring.MetricsSupport
+import javax.management.{JMX, ObjectName}
 import com.advancedtelematic.libats.slick.db.DatabaseConfig
+import com.advancedtelematic.metrics.MetricsSupport
 import com.codahale.metrics.{Gauge, Metric, MetricSet}
 import slick.jdbc.hikaricp.HikariCPJdbcDataSource
 import slick.util.AsyncExecutorMXBean
